chore: Add time delay in ReturnNodeValue __call__ method
The code changes include adding a time delay of 1 second in the  method of the  class. This delay is introduced using the  function. The purpose of this delay is to simulate a time-consuming operation before printing the value being added to the state's aggregate. This delay can be useful for testing and debugging purposes.

Note: The recent user commits and repository commits do not provide any additional context for generating the commit message.